common-algorithm/src/test/java/com/dwarveswork/algorithm/ScryptTest.java
/*
 * SPDX-License-Identifier: MIT
 * Copyright (c) 2020 dwarveswork
 */

package com.dwarveswork.algorithm;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.nio.charset.StandardCharsets;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.ValueSource;

class ScryptTest {

  @ParameterizedTest
  @ValueSource(strings = {
      "::4:1:1:64:77d6576238657b203b19ca42c18a0497f16b4844e3074ae8dfdffa3fede21442fcd0069ded0948f8326a753a0fc81f17e8d3e0fb2e0d3628cf35e20c38d18906",
      "password:NaCl:10:8:16:64:fdbabe1c9d3472007856e7190d01e9fe7c6ad7cbc8237830e77376634b3731622eaf30d92e22a3886ff109279d9830dac727afb94a83ee6d8360cbdfa2cc0640",
      "pleaseletmein:SodiumChloride:14:8:1:64:7023bdcb3afd7348461c06cd81fd38ebfda8fbba904f8e3ea9b543f6545da1f2d5432955613f0fcf62d49705242a9af9e61e85dc0d651e40dfcf017b45575887",
      "pleaseletmein:SodiumChloride:20:8:1:64:2101cb9b6a511aaeaddbbe09cf70f881ec568d574a2ffd4dabe5ee9820adaa478e56fd8f4ba5d09ffa1c6d927c40f4c337304049e8a952fbcbf45c6fa77a41a4"
  })
  void testEncode(String value) {
    String[] parts = value.split(":");
    String hex = Scrypt.encodeToHex(parts[0].getBytes(StandardCharsets.UTF_8), parts[1].getBytes(StandardCharsets.UTF_8),
                                    Integer.parseInt(parts[2]), Integer.parseInt(parts[3]), Integer.parseInt(parts[4]), Integer.parseInt(parts[5]));
    assertEquals(hex, parts[6]);
  }
}